requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.1.0.post1
Werkzeug==2.3.7
python-dotenv==1.0.0
bleach==6.0.0
python-socketio==5.8.0
python-engineio==4.7.1
eventlet==0.33.3
